model: move various validation from roles.go to new validation.go (NFC)
Just moving some code around to make giant file somewhat less giant
